==> notebooks/4_train_test_id_csv.ipynb <==
%matplotlib inline
import numpy as np
from skimage import io
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.utils import shuffle
import os

OUTPUT_DIR = '../data/csv/'
PREPROCCESSED_DATA_DIR = '../data/preprocess_data'

def get_frame_id(fname):
    return fname.split("--")[1].split('.')[0]

def get_track_id(fname):
    return fname.split("--")[0].split('bee')[1]

train_path = os.path.join(PREPROCCESSED_DATA_DIR, 'train_data', 'id')
train_files = os.listdir(train_path)

test_path = os.path.join(PREPROCCESSED_DATA_DIR, 'test_data', 'id')
test_files = os.listdir(test_path)

# train_files

# test_files

ids = [get_track_id(i) for i in train_files]
ids = list(np.unique(ids))

ids

train_ids = list()

for fname in train_files:
    train_ids.append(ids.index(get_track_id(fname)))

# train_ids

test_ids = list()

for fname in test_files:
    test_ids.append(ids.index(get_track_id(fname)))

# test_ids

train_files = [os.path.join(train_path, fname) for fname in train_files]

# train_files

test_files = [os.path.join(test_path, fname) for fname in test_files]

# test_files

train_df_dict = {'X_train': train_files, 'y':train_ids}
train_df = pd.DataFrame(train_df_dict)

train_df.to_csv(os.path.join(OUTPUT_DIR, 'train_id.csv'), index=False)

# train_df

test_df_dict = {'X_test': test_files, 'y':test_ids}
test_df = pd.DataFrame(test_df_dict)

test_df.to_csv(os.path.join(OUTPUT_DIR, 'test_id.csv'), index=False)

test_df
