==> notebooks/5_siamese_network.ipynb <==
# Siamese Network

#### dependecies

%matplotlib inline
import random
from skimage import io
import matplotlib.pyplot as plt
from ipywidgets import interact
from keras.models import Model, Sequential 
from keras.models import load_model
from keras.callbacks import ModelCheckpoint, EarlyStopping, TensorBoard
from keras.applications.vgg16 import VGG16
from keras.optimizers import SGD
import numpy as np
from keras.layers import Input, Flatten, Dense, Dropout, Conv2D, MaxPooling2D, Merge
from keras.layers.normalization import BatchNormalization
from keras.optimizers import RMSprop
from keras import backend as K
from keras.applications.vgg16 import VGG16, preprocess_input, decode_predictions
from keras.preprocessing.image import img_to_array
from keras.utils import to_categorical
from keras.layers.core import Lambda
import keras
import os
from sklearn.metrics import accuracy_score
import seaborn as sns
import pandas as pd
import sys
sys.path.append("../")
from networks.networks import *

CSV_DIR = '../data/csv/'
MODEL_DIR = '../models/'
LOGS_DIR = '../logs'

### Load Data

train_df = pd.read_csv(os.path.join(CSV_DIR, 'train_pairs.csv'))
test_df = pd.read_csv(os.path.join(CSV_DIR, 'test_pairs.csv'))

# train_df.head(20)

X1_paths = list(train_df.X1[0:20])
X2_paths = list(train_df.X2[0:20])

X1_test_paths = list(test_df.X1[0:10])
X2_test_paths = list(test_df.X2[0:10])

X1 = np.array(io.imread_collection(X1_paths))
X2 = np.array(io.imread_collection(X2_paths))

X1_test = np.array(io.imread_collection(X1_test_paths))
X2_test = np.array(io.imread_collection(X2_test_paths))

y = np.array(train_df.y)
y_test = np.array(test_df.y)

X1.shape, X2.shape, y.shape

X1_test.shape, X2_test.shape, y_test.shape

#### Loss and distance related Functions

def euclidean_distance(vects):
    x, y = vects
    return K.sqrt(K.maximum(K.mean(K.square(x - y), axis=1, keepdims=True), K.epsilon()))


def eucl_dist_output_shape(shapes):
    shape1, shape2 = shapes
    return (shape1[0], 1)

def accuracy(y_true, y_pred):
    '''Compute classification accuracy with a fixed threshold on distances.
    '''
    threshold = 0.2
    return 1 -  K.mean(K.equal(y_true, K.cast(y_pred < threshold, y_true.dtype)))


def accuracy5(y_true, y_pred):
    '''Compute classification accuracy with a fixed threshold on distances.
    '''
    threshold = 0.5
    return 1 -  K.mean(K.equal(y_true, K.cast(y_pred < threshold, y_true.dtype)))

def accuracy1(y_true, y_pred):
    '''Compute classification accuracy with a fixed threshold on distances.
    '''
    threshold = 1.0
    return 1 -  K.mean(K.equal(y_true, K.cast(y_pred < threshold, y_true.dtype)))


def contrastive_loss(y_diff, y_dist):
    '''Contrastive loss from Hadsell-et-al.'06
    http://yann.lecun.com/exdb/publis/pdf/hadsell-chopra-lecun-06.pdf
    '''
    margin = 1.0
    return K.mean((1 - y_diff) * K.square(y_dist) +
                  (y_diff) * K.square(K.maximum(margin - y_dist, 0)))


### Build Network Architecture

network = base_network('vggnet_6')

network.summary()

### Configure Siamese Network

input_a = Input(shape=[230, 105, 3])
input_b = Input(shape=[230, 105, 3])

processed_a = network(input_a)
processed_b = network(input_b)

#### Merge networks with euclidean distance

distance = Lambda(euclidean_distance,
                          output_shape=eucl_dist_output_shape)([processed_a, processed_b])

model = Model([input_a, input_b], distance)

# model.save(os.path.join(MODEL_DIR, "train_4.h5"))

#### Configure loss and other parameters


model.compile(loss=contrastive_loss, optimizer='SGD', metrics=[accuracy, accuracy1, accuracy5])

model.summary()

#### Set-up TensorBoard

# tf_board = TensorBoard(os.path.join(LOGS_DIR, 'logs', 'exp_5'))

earlystop = EarlyStopping(monitor='val_acc', min_delta=0.0001, patience=5, verbose=1, mode='auto')

### Train Siamese Network 

history = model.fit([X1, X2], y[0:20], 
                    batch_size=1, epochs=10, verbose=1, callbacks=[earlystop],
                    validation_data=([X1_test, X2_test], y_test[0:10]))

history.history

# model.save("../models/exp_5.h5")
# model.save_weights("../weights/exp_5-w.h5")

model.predict([X1_test, X2_test]), y_test[0:10]
